Update to .NET 9 Update of related packages Cleanup of class structure Minor breaking change in Subscription types - Enum was renamed to adhere to naming schema of file Bumped up version to 2.2.0
