+++ Twitch EventSub library/API/Enums/ConditionTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace Twitch.EventSub.API.Models
+﻿namespace Twitch.EventSub.API.Enums
 {
     public enum ConditionTypes
     {
