+++ Twitch EventSub library/Messages/NotificationMessage/Events/Automod/AutomodMessageUpdateEvent.cs
@@ -1,6 +1,5 @@
 ﻿using Newtonsoft.Json;
-using Twitch.EventSub.Messages.NotificationMessage.Events.ChannelChat;
-using Twitch.EventSub.Messages.NotificationMessage.Events.ChannelSubscription;
+using Twitch.EventSub.Messages.NotificationMessage.Events.Automod.Models;
 
 namespace Twitch.EventSub.Messages.NotificationMessage.Events.Automod
 {
@@ -44,14 +43,5 @@ namespace Twitch.EventSub.Messages.NotificationMessage.Events.Automod
 
         [JsonProperty("fragments")]
         public Fragments Fragments { get; set; }
-    }
-
-    public class Fragments
-    {
-        [JsonProperty("emotes")]
-        public List<Emote> Emotes { get; set; }
-
-        [JsonProperty("cheermotes")]
-        public List<Cheermote> Cheermotes { get; set; }
     }
 }
