+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelChat/Models/BitsBadgeTierNotification.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelChat.Models
+{
+    public class BitsBadgeTierNotification
+    {
+        [JsonProperty("tier")]
+        public int Tier { get; set; }
+    }
+}
