+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelChat/Models/MentionNotification.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelChat.Models
+{
+    public class MentionNotification
+    {
+        [JsonProperty("user_id")]
+        public string UserId { get; set; }
+
+        [JsonProperty("user_name")]
+        public string UserName { get; set; }
+
+        [JsonProperty("user_login")]
+        public string UserLogin { get; set; }
+    }
+}
