+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelChat/Models/PrimePaidUpgradeNotification.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelChat.Models
+{
+    public class PrimePaidUpgradeNotification
+    {
+        [JsonProperty("sub_tier")]
+        public string SubTier { get; set; }
+    }
+}
