+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelChat/Models/RaidNotification.cs
@@ -1,0 +1,22 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelChat.Models
+{
+    public class RaidNotification
+    {
+        [JsonProperty("user_id")]
+        public string UserId { get; set; }
+
+        [JsonProperty("user_name")]
+        public string UserName { get; set; }
+
+        [JsonProperty("user_login")]
+        public string UserLogin { get; set; }
+
+        [JsonProperty("viewer_count")]
+        public int ViewerCount { get; set; }
+
+        [JsonProperty("profile_image_url")]
+        public string ProfileImageUrl { get; set; }
+    }
+}
