+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelPoints/Models/Image.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelPoints.Models
+{
+    public class Image
+    {
+        [JsonProperty("url_1x")]
+        public string Url1x { get; set; }
+
+        [JsonProperty("url_2x")]
+        public string Url2x { get; set; }
+
+        [JsonProperty("url_4x")]
+        public string Url4x { get; set; }
+    }
+}
