+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelPoints/Models/MaxPerUserPerStream.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelPoints.Models
+{
+    public class MaxPerUserPerStream
+    {
+        [JsonProperty("is_enabled")]
+        public bool IsEnabled { get; set; }
+
+        [JsonProperty("value")]
+        public int Value { get; set; }
+    }
+}
