+++ Twitch EventSub library/Messages/NotificationMessage/Events/ChannelPoints/Models/Reward.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace Twitch.EventSub.Messages.NotificationMessage.Events.ChannelPoints.Models
+{
+    public class Reward
+    {
+        [JsonProperty("id")]
+        public string Id { get; set; }
+
+        [JsonProperty("title")]
+        public string Title { get; set; }
+
+        [JsonProperty("cost")]
+        public int Cost { get; set; }
+
+        [JsonProperty("prompt")]
+        public string Prompt { get; set; }
+    }
+}
