+++ Twitch EventSub library/SubsRegister/Models/RegisterItem.cs
@@ -1,0 +1,13 @@
+﻿using Twitch.EventSub.API.Enums;
+
+namespace Twitch.EventSub.SubsRegister.Models
+{
+    public class RegisterItem
+    {
+        public string Key { get; set; }
+        public Type SpecificObject { get; set; }
+        public SubscriptionTypes SubscriptionType { get; set; }
+        public string Ver { get; set; }
+        public List<ConditionTypes> Conditions { get; set; }
+    }
+};
