+++ Twitch EventSub library/User/UserSequencer.cs
@@ -206,8 +206,7 @@ namespace Twitch.EventSub.User
                 default:
                     _logger.LogError("[OnRefreshTokenRequestAsync] Unexpected state: {State}", StateMachine.State);
                     throw new InvalidOperationException("[EventSubClient] - [UserSequencer] OnRefreshTokenRequestAsync went into unknown state");
-            };
-        }
+            }
 
         /// <summary>
         /// Awaits the welcome message from the WebSocket.
