+++ TwitchEventSub_Websocket.Tests/SanityChecks/RegistryTests.cs
@@ -1,9 +1,9 @@
 ﻿using System.Reflection;
-using Twitch.EventSub;
+using Twitch.EventSub.API.Enums;
 using Twitch.EventSub.API.Extensions;
-using Twitch.EventSub.API.Models;
-using Twitch.EventSub.User;
 using Twitch.EventSub.SubsRegister;
+using Twitch.EventSub.SubsRegister.Models;
+using Twitch.EventSub.User;
 
 
 namespace TwitchEventSub_Websocket.Tests.SanityChecks
@@ -81,7 +81,7 @@ namespace TwitchEventSub_Websocket.Tests.SanityChecks
             // Retrieve all subscription types from TypeVersionConditionMap
             var subscriptionTypes = typeof(CreateSubscriptionRequestExtension)
                 .GetField("TypeVersionConditionMap", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Static)
-                ?.GetValue(null) as Dictionary<SubscriptionType, (string Type, string Version, List<ConditionTypes> Conditions)>;
+                ?.GetValue(null) as Dictionary<SubscriptionTypes, (string Type, string Version, List<ConditionTypes> Conditions)>;
 
             // Ensure registryItems and subscriptionTypes are not null
             Assert.NotNull(registryItems);
